# apps/api-service/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MCPFactory API Service",
    "description": "API Gateway for MCPFactory. Handles authentication, proxies to internal services, and exposes the public REST API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.mcpfactory.org"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer JWT from Clerk (dashboard)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for MCP clients and service-to-service communication"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "BrandScrapeRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "description": "Brand website URL to scrape"
          },
          "skipCache": {
            "type": "boolean",
            "description": "Skip cached results and force re-scrape"
          }
        },
        "required": [
          "url"
        ]
      },
      "IcpSuggestionRequest": {
        "type": "object",
        "properties": {
          "brandUrl": {
            "type": "string",
            "minLength": 1,
            "description": "Brand website URL"
          }
        },
        "required": [
          "brandUrl"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/brand/scrape": {
      "post": {
        "tags": [
          "Brand"
        ],
        "summary": "Scrape brand info",
        "description": "Scrape brand information from a URL using the scraping service",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scraped brand information"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brand/by-url": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get brand by URL",
        "description": "Get cached brand info by website URL",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand website URL"
            },
            "required": true,
            "description": "Brand website URL",
            "name": "url",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cached brand information"
          },
          "400": {
            "description": "Missing url param",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brands": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "List brands",
        "description": "Get all brands for the organization",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of brands"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brands/{id}": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get a brand",
        "description": "Get a single brand by ID",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand ID"
            },
            "required": true,
            "description": "Brand ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand data"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brands/{id}/sales-profile": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get brand sales profile",
        "description": "Get the sales profile for a specific brand",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand ID"
            },
            "required": true,
            "description": "Brand ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand sales profile"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Sales profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brand/sales-profiles": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "List sales profiles",
        "description": "Get all sales profiles (brands) for the organization",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of sales profiles"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brand/icp-suggestion": {
      "post": {
        "tags": [
          "Brand"
        ],
        "summary": "Get ICP suggestion",
        "description": "Get AI-generated Ideal Customer Profile suggestion (Apollo-compatible search params) for a brand URL",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IcpSuggestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ICP suggestion (Apollo-compatible search params)"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brands/{id}/cost-breakdown": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get brand cost breakdown",
        "description": "Get cost breakdown by cost name for all runs associated with a brand, from runs-service",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand ID"
            },
            "required": true,
            "description": "Brand ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Cost breakdown by cost name"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brands/{id}/runs": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get brand runs",
        "description": "Get extraction runs for a brand (sales-profile, icp-extraction) enriched with cost data",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Brand ID"
            },
            "required": true,
            "description": "Brand ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand extraction runs with cost data"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/brand/{id}": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get brand scrape result",
        "description": "Get brand scrape result by scrape ID",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Scrape ID"
            },
            "required": true,
            "description": "Scrape ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Brand scrape result"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}